The WebSocket module becomes a directory split along its seams
A pure move: detection and the extension-negotiation surface stay in
proxy/websocket/mod.rs, the upgrade handshake moves to handshake.rs, and
the post-101 frame relay to relay.rs, each with the tests whose subjects
moved. No behavior changes; the citations index follows the cites to
their new paths.
