--- lint-http-proxy/src/proxy/websocket.rs
@@ -1,1705 +1,0 @@
-// SPDX-FileCopyrightText: 2026 Alexandre Gomes Gaigalas <alganet@gmail.com>
-//
-// SPDX-License-Identifier: ISC
-
-//! WebSocket upgrade handshake and bidirectional frame relay.
-
-use bytes::Bytes;
-use http_body_util::{BodyExt, Full};
-use hyper::{Method, Request, Response, Uri};
-use hyper_util::rt::TokioIo;
-use std::convert::Infallible;
-use std::sync::Arc;
-use tokio::time::Instant;
-use tokio_util::sync::CancellationToken;
-use tracing::{debug, error};
-
-use crate::capture::CaptureWriter;
-
-use super::exchange::record_error_transaction;
-use super::hop_by_hop::{format_http_version, parse_connection_tokens};
-use super::pipeline::ProtocolEventPipeline;
-use super::{boxed_full, BoxError, ResponseBody, Shared};
-
-/// Check if a request is a WebSocket upgrade request.
-///
-/// A WebSocket handshake requires `Connection: Upgrade` (as a distinct token,
-/// possibly alongside others) and `Upgrade: websocket`.
-///
-/// Both sentences say "include", not "equal", and that is the whole reason the
-/// two checks are shaped differently. Connection is a token list, so `include`
-/// means membership -- `Connection: keep-alive, Upgrade` is a valid handshake
-/// and a string compare would reject it, which is why this asks
-/// `parse_connection_tokens` rather than looking at the field value. The
-/// keyword and the token are also matched case-insensitively, which the
-/// document establishes elsewhere for each field rather than here.
-///
-// cite(RFC 6455 § 4.1): "The request MUST contain an |Upgrade| header field whose value MUST include the "websocket" keyword."
-// cite(RFC 6455 § 4.1): "The request MUST contain a |Connection| header field whose value MUST include the "Upgrade" token."
-pub(super) fn is_websocket_upgrade<B>(req: &Request<B>) -> bool {
-    let connection_tokens = parse_connection_tokens(req.headers().get(hyper::header::CONNECTION));
-    let has_upgrade = connection_tokens.contains("upgrade");
-    let is_websocket = req
-        .headers()
-        .get(hyper::header::UPGRADE)
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.eq_ignore_ascii_case("websocket"))
-        .unwrap_or(false);
-    has_upgrade && is_websocket
-}
-
-/// Handle a WebSocket upgrade request: connect directly to upstream, relay
-/// frames via tokio-tungstenite, and capture the session.
-#[allow(clippy::too_many_arguments)]
-pub(super) async fn handle_websocket_upgrade(
-    upstream_req: Request<Full<Bytes>>,
-    client_on_upgrade: hyper::upgrade::OnUpgrade,
-    uri: &Uri,
-    scheme: &hyper::http::uri::Scheme,
-    started: &Instant,
-    client_id: &crate::state::ClientIdentifier,
-    method: &Method,
-    uri_str: &str,
-    req_headers: &hyper::HeaderMap,
-    req_version: &str,
-    body_bytes: Bytes,
-    req_trailers: Option<hyper::HeaderMap>,
-    shared: Arc<Shared>,
-    conn_metadata: Arc<crate::connection::ConnectionMetadata>,
-) -> Result<Response<ResponseBody>, Infallible> {
-    // Connect directly to upstream with upgrade support, reusing the shared
-    // outbound TLS config (loaded once at startup).
-    let (mut sender, _conn_handle) =
-        match connect_upstream_for_upgrade(uri, scheme, &shared.upstream.tls_config).await {
-            Ok(s) => s,
-            Err(e) => {
-                error!("websocket upstream connect error: {}", e);
-                record_handshake_failure(
-                    &shared,
-                    client_id,
-                    method,
-                    uri_str,
-                    req_headers,
-                    req_version,
-                    &body_bytes,
-                    &conn_metadata,
-                    started,
-                )
-                .await;
-                return Ok(upstream_error_response(&e));
-            }
-        };
-
-    // Send the upgrade request to the upstream server
-    let mut upstream_resp = match sender.send_request(upstream_req).await {
-        Ok(r) => r,
-        Err(e) => {
-            error!("websocket upstream request error: {}", e);
-            record_handshake_failure(
-                &shared,
-                client_id,
-                method,
-                uri_str,
-                req_headers,
-                req_version,
-                &body_bytes,
-                &conn_metadata,
-                started,
-            )
-            .await;
-            return Ok(upstream_error_response(&e));
-        }
-    };
-
-    let status = upstream_resp.status().as_u16();
-    let headers = upstream_resp.headers().clone();
-    let resp_ver = format_http_version(upstream_resp.version());
-    let duration = started.elapsed().as_millis() as u64;
-
-    // Record the HTTP transaction (the 101 handshake)
-    let mut tx = crate::http_transaction::HttpTransaction::new(
-        client_id.clone(),
-        method.as_str().to_string(),
-        uri_str.to_string(),
-    );
-    tx.request.headers = req_headers.clone();
-    tx.request.version = req_version.to_string();
-    tx.request.body_length = Some(body_bytes.len() as u64);
-    tx.request.trailers = req_trailers;
-    tx.request_body = Some(body_bytes);
-    tx.response = Some(crate::http_transaction::ResponseInfo {
-        status,
-        version: resp_ver,
-        headers: headers.clone(),
-        // A 101 has no body; a non-101 response body streams through to the
-        // client but is not captured here, so record it as unknown rather than
-        // falsely claiming zero length.
-        //
-        // The `Some(0)` is knowledge, not a guess -- the sentence below is what
-        // makes zero the only possible answer for a 101, and it is why the two
-        // arms are asymmetric. `None` here means "we did not look"; `Some(0)`
-        // means "there is nothing to look at". Only one of those can be said
-        // without reading the body, and only for 1xx.
-        //
-        // cite(RFC 9110 § 15.2): "A 1xx response is terminated by the end of the header section; it cannot contain content or trailers."
-        body_length: if status == 101 { Some(0) } else { None },
-        trailers: None,
-    });
-    tx.timing = crate::http_transaction::TimingInfo {
-        duration_ms: duration,
-    };
-    tx.connection_id = Some(conn_metadata.id);
-    tx.sequence_number = Some(conn_metadata.next_sequence_number());
-    if status == 101 {
-        tx.was_upgraded = true;
-        tx.upgrade_protocol = headers
-            .get("upgrade")
-            .and_then(|v| v.to_str().ok())
-            .map(|s| s.to_string());
-    }
-
-    let tx_id = tx.id;
-
-    if status == 101 {
-        shared.pipeline().commit(tx).await;
-
-        // Extract the server-side upgraded IO
-        let server_upgraded = hyper::upgrade::on(&mut upstream_resp);
-
-        // Build the 101 response to send back to the client.
-        // Forward ALL headers including upgrade-related ones (Connection, Upgrade,
-        // Sec-WebSocket-Accept) — do NOT strip hop-by-hop headers for 101. The one
-        // exception is the extension negotiation, which stops at the relay:
-        // see `without_extension_negotiation`. The capture above already
-        // recorded the 101 as received.
-        let mut resp_builder = Response::builder().status(101);
-        for (name, value) in without_extension_negotiation(&headers).iter() {
-            resp_builder = resp_builder.header(name, value);
-        }
-        let resp = resp_builder
-            .body(boxed_full(Bytes::new()))
-            .unwrap_or_else(|_| Response::new(boxed_full(Bytes::new())));
-
-        // Spawn the background relay, holding a connection permit and a shutdown
-        // token for its lifetime: the permit counts the live session against
-        // `max_connections` (and makes the drain barrier wait for it), the token
-        // lets it close promptly on shutdown. The permit is best-effort —
-        // an already-upgraded connection can't be rejected if we're at capacity.
-        let captures_clone = shared.captures.clone();
-        let connection_id = conn_metadata.id;
-        let pe_pipeline = shared.protocol_event_pipeline();
-        let relay_permit = shared.semaphore.clone().try_acquire_owned().ok();
-        if relay_permit.is_none() {
-            debug!("websocket relay starting without a connection permit (at capacity)");
-        }
-        let relay_shutdown = shared.shutdown.clone();
-        let negotiated = accepted_extensions(&headers);
-        tokio::spawn(async move {
-            let _relay_permit = relay_permit;
-            // Wait for both sides to complete the upgrade
-            let (client_io, server_io) = match tokio::try_join!(client_on_upgrade, server_upgraded)
-            {
-                Ok((c, s)) => (c, s),
-                Err(e) => {
-                    error!("websocket upgrade failed: {}", e);
-                    return;
-                }
-            };
-
-            relay_websocket(
-                TokioIo::new(client_io),
-                TokioIo::new(server_io),
-                tx_id,
-                captures_clone,
-                connection_id,
-                negotiated,
-                pe_pipeline,
-                relay_shutdown,
-            )
-            .await;
-        });
-
-        Ok(resp)
-    } else {
-        // Upstream did not accept the upgrade: it is a normal HTTP response.
-        // Stream it back to the client (no buffering, no over-limit 502) and
-        // commit the handshake transaction. The non-101 response body is not
-        // separately captured here (as before — only the handshake metadata is
-        // recorded), so a plain streaming forward suffices.
-        shared.pipeline().commit(tx).await;
-
-        let inner = upstream_resp
-            .into_body()
-            .map_err(|e| -> BoxError { e.into() })
-            .boxed_unsync();
-        let mut resp_builder = Response::builder().status(status);
-        for (name, value) in super::exchange::filter_response_headers(&headers, status).iter() {
-            resp_builder = resp_builder.header(name, value);
-        }
-        let resp = resp_builder
-            .body(inner)
-            .unwrap_or_else(|_| Response::new(boxed_full(Bytes::new())));
-
-        Ok(resp)
-    }
-}
-
-/// Build the 502 returned to the client when a WebSocket upstream handshake
-/// fails (connect or request-send error).
-fn upstream_error_response(e: impl std::fmt::Display) -> Response<ResponseBody> {
-    let body = boxed_full(Bytes::from(format!("websocket upstream error: {}", e)));
-    Response::builder()
-        .status(502)
-        .body(body)
-        .unwrap_or_else(|_| Response::new(boxed_full(Bytes::from("upstream error"))))
-}
-
-/// Record a transaction for a WebSocket handshake that failed before the
-/// upstream produced any response, so the request is not silently lost. Routes
-/// through the pipeline (lint → state → capture) and consumes one sequence
-/// number, matching the success path.
-#[allow(clippy::too_many_arguments)]
-async fn record_handshake_failure(
-    shared: &Arc<Shared>,
-    client_id: &crate::state::ClientIdentifier,
-    method: &Method,
-    uri_str: &str,
-    req_headers: &hyper::HeaderMap,
-    req_version: &str,
-    body_bytes: &Bytes,
-    conn_metadata: &crate::connection::ConnectionMetadata,
-    started: &Instant,
-) {
-    let duration = started.elapsed().as_millis() as u64;
-    record_error_transaction(
-        shared,
-        client_id,
-        method.as_str(),
-        uri_str,
-        req_headers,
-        req_version,
-        502,
-        None,
-        duration,
-        Some(body_bytes.clone()),
-        conn_metadata.id,
-        conn_metadata.next_sequence_number(),
-        false,
-        false,
-    )
-    .await;
-}
-
-/// Open a direct TCP (or TLS) connection to the upstream host and perform
-/// an HTTP/1.1 handshake with upgrade support.
-///
-/// `fallback_scheme` is used only when the URI itself has no scheme set
-/// (origin-form requests).  An absolute-form URI's own scheme always wins,
-/// so the scheme used for TLS and the default port can never disagree with
-/// what's in the URI.
-async fn connect_upstream_for_upgrade(
-    uri: &Uri,
-    fallback_scheme: &hyper::http::uri::Scheme,
-    tls_config: &Arc<rustls::ClientConfig>,
-) -> anyhow::Result<(
-    hyper::client::conn::http1::SendRequest<Full<Bytes>>,
-    tokio::task::JoinHandle<Result<(), hyper::Error>>,
-)> {
-    let host = uri
-        .host()
-        .ok_or_else(|| anyhow::anyhow!("missing host in URI"))?;
-    let scheme = uri.scheme().unwrap_or(fallback_scheme);
-    let is_https = *scheme == hyper::http::uri::Scheme::HTTPS;
-    let port = uri.port_u16().unwrap_or(if is_https { 443 } else { 80 });
-
-    let tcp = tokio::net::TcpStream::connect((host, port)).await?;
-
-    if is_https {
-        // The trust store was loaded once at startup; reuse the shared config
-        // (an `Arc` bump) instead of re-reading native certs per upgrade.
-        let connector = tokio_rustls::TlsConnector::from(tls_config.clone());
-        let server_name = rustls::pki_types::ServerName::try_from(host.to_string())?;
-        let tls_stream = connector.connect(server_name, tcp).await?;
-
-        let (sender, conn) =
-            hyper::client::conn::http1::handshake(TokioIo::new(tls_stream)).await?;
-        let handle = tokio::spawn(conn.with_upgrades());
-        Ok((sender, handle))
-    } else {
-        let (sender, conn) = hyper::client::conn::http1::handshake(TokioIo::new(tcp)).await?;
-        let handle = tokio::spawn(conn.with_upgrades());
-        Ok((sender, handle))
-    }
-}
-
-/// The message minus its `Sec-WebSocket-Extensions` field: the extension
-/// negotiation stops at the relay, in both directions.
-///
-/// **The relay is a WebSocket endpoint on each leg, not a tunnel.** Every frame
-/// passes through tungstenite, which implements no extension and returns
-/// `ProtocolError::NonZeroReservedBits` for any frame that sets one — so a
-/// negotiation relayed end-to-end authorises frames the middle cannot read.
-/// That was live: the client's `permessage-deflate` offer reached the origin,
-/// the origin's acceptance reached the client, and the first compressed frame
-/// killed the session (RULECITES P48). An intermediary that cannot read an
-/// extension's frames must not relay its negotiation, so the offer is removed
-/// from the upstream handshake request and the acceptance — which a conforming
-/// origin then never sends, but a broken one still might — from the `101`
-/// forwarded to the client.
-///
-/// **Captures are untouched.** Both the request and the `101` are recorded as
-/// received, offer and acceptance included, before this runs — the rules read
-/// what each party wrote, not what the relay could live with.
-// cite(RFC 6455 § 9.1): "Note that like other HTTP header fields, this header field MAY be split or combined across multiple lines."
-pub(super) fn without_extension_negotiation(headers: &hyper::HeaderMap) -> hyper::HeaderMap {
-    let mut out = headers.clone();
-    // `HeaderMap::remove` takes every value of the name, so a field split
-    // across lines — which § 9.1 permits in as many words — leaves nothing.
-    out.remove("sec-websocket-extensions");
-    out
-}
-
-/// What a `101` settled about extensions, read from the response and not from
-/// the offer.
-///
-/// **Presence is the whole decision, and that is why the value is not parsed
-/// here.** RFC 6455 § 9.1 makes the server's list the extensions in use, and a
-/// client's own field only an offer it may not act on — so a `101` with no such
-/// field is the connection's answer that nothing was accepted, which is what
-/// lets a frame rule read the reserved bits and opcodes at all. A `101` that
-/// carries the field stands those findings down whatever it holds: which
-/// extension defines which bit is that extension's document's business.
-///
-/// **An unreadable value is still a field.** Reading it through a UTF-8 decoder
-/// would turn a `Sec-WebSocket-Extensions` carrying `obs-text` into a handshake
-/// that accepted nothing — a claim about the exchange, and the one direction
-/// that *licenses* findings. The shared as-written reader gives one `char` per
-/// octet and joins the lines the way § 5.2 does, so the record holds what the
-/// server wrote.
-///
-/// The absent case is `NoneAccepted` and never `Unrecorded`: this proxy watched
-/// the handshake, so it knows a difference a capture written elsewhere cannot
-/// state.
-///
-// cite(RFC 6455 § 9.1): "The extensions listed by the server in response represent the extensions actually in use for the connection."
-fn accepted_extensions(headers: &hyper::HeaderMap) -> crate::protocol_event::NegotiatedExtensions {
-    use crate::protocol_event::NegotiatedExtensions;
-    match crate::helpers::headers::combined_field_value_as_written(
-        headers,
-        "sec-websocket-extensions",
-    ) {
-        Some(value) => NegotiatedExtensions::Accepted(value),
-        None => NegotiatedExtensions::NoneAccepted,
-    }
-}
-
-/// Build the protocol event for a single relayed WebSocket frame, stamped with
-/// its arrival time. Thin wrapper over the shared
-/// [`WebSocketMessageInfo::frame_event`] mapping so live and offline replay
-/// can't drift.
-fn ws_frame_event(
-    connection_id: uuid::Uuid,
-    session_id: uuid::Uuid,
-    info: &crate::websocket_session::WebSocketMessageInfo,
-    extensions: &crate::protocol_event::NegotiatedExtensions,
-) -> crate::protocol_event::ProtocolEvent {
-    info.frame_event(chrono::Utc::now(), connection_id, session_id, extensions)
-}
-
-/// Relay WebSocket messages between client and server, recording each message
-/// for capture. Uses tokio-tungstenite for proper RFC 6455 frame parsing.
-///
-/// The eighth parameter is what the handshake settled about extensions, and it
-/// is passed rather than re-read because the `101` this session came from is
-/// gone by the time the relay runs — the same reason `tx_id` is passed. Its
-/// caller carries the same allow, for the same reason.
-#[allow(clippy::too_many_arguments)]
-async fn relay_websocket(
-    client_io: impl tokio::io::AsyncRead + tokio::io::AsyncWrite + Unpin + Send,
-    server_io: impl tokio::io::AsyncRead + tokio::io::AsyncWrite + Unpin + Send,
-    tx_id: uuid::Uuid,
-    captures: CaptureWriter,
-    connection_id: uuid::Uuid,
-    extensions: crate::protocol_event::NegotiatedExtensions,
-    pipeline: ProtocolEventPipeline,
-    shutdown: CancellationToken,
-) {
-    use crate::websocket_session::{MessageDirection, WebSocketMessageInfo, WebSocketSession};
-    use futures_util::{SinkExt, StreamExt};
-    use tokio_tungstenite::tungstenite::protocol::Role;
-
-    let client_ws =
-        tokio_tungstenite::WebSocketStream::from_raw_socket(client_io, Role::Server, None).await;
-    let server_ws =
-        tokio_tungstenite::WebSocketStream::from_raw_socket(server_io, Role::Client, None).await;
-
-    let (mut client_write, mut client_read) = client_ws.split();
-    let (mut server_write, mut server_read) = server_ws.split();
-
-    let session_id = uuid::Uuid::new_v4();
-    let messages = Arc::new(tokio::sync::Mutex::new(Vec::<WebSocketMessageInfo>::new()));
-    let violations = Arc::new(tokio::sync::Mutex::new(Vec::<crate::lint::Violation>::new()));
-    let close_code = Arc::new(tokio::sync::Mutex::new(None::<u16>));
-    let start = Instant::now();
-
-    let ext_c2s = extensions.clone();
-    let ext_s2c = extensions.clone();
-    let msgs_c2s = messages.clone();
-    let viols_c2s = violations.clone();
-    let close_c2s = close_code.clone();
-    let pipe_c2s = pipeline.clone();
-    let c2s = async move {
-        while let Some(result) = client_read.next().await {
-            match result {
-                Ok(msg) => {
-                    let info = message_to_info(&msg, MessageDirection::Client);
-                    if let tokio_tungstenite::tungstenite::Message::Close(Some(ref frame)) = msg {
-                        let mut cc = close_c2s.lock().await;
-                        if cc.is_none() {
-                            *cc = Some(frame.code.into());
-                        }
-                    }
-                    // Emit protocol event and lint it
-                    let pe = ws_frame_event(connection_id, session_id, &info, &ext_c2s);
-                    let v = pipe_c2s.commit(&pe);
-                    if !v.is_empty() {
-                        viols_c2s.lock().await.extend(v);
-                    }
-
-                    msgs_c2s.lock().await.push(info);
-                    if server_write.send(msg).await.is_err() {
-                        break;
-                    }
-                }
-                Err(_) => break,
-            }
-        }
-    };
-
-    let msgs_s2c = messages.clone();
-    let viols_s2c = violations.clone();
-    let close_s2c = close_code.clone();
-    let pipe_s2c = pipeline;
-    let s2c = async move {
-        while let Some(result) = server_read.next().await {
-            match result {
-                Ok(msg) => {
-                    let info = message_to_info(&msg, MessageDirection::Server);
-                    if let tokio_tungstenite::tungstenite::Message::Close(Some(ref frame)) = msg {
-                        let mut cc = close_s2c.lock().await;
-                        if cc.is_none() {
-                            *cc = Some(frame.code.into());
-                        }
-                    }
-                    // Emit protocol event and lint it
-                    let pe = ws_frame_event(connection_id, session_id, &info, &ext_s2c);
-                    let v = pipe_s2c.commit(&pe);
-                    if !v.is_empty() {
-                        viols_s2c.lock().await.extend(v);
-                    }
-
-                    msgs_s2c.lock().await.push(info);
-                    if client_write.send(msg).await.is_err() {
-                        break;
-                    }
-                }
-                Err(_) => break,
-            }
-        }
-    };
-
-    // Run both directions concurrently; when either finishes, the session is
-    // done. On shutdown, break promptly (dropping the IO halves closes both
-    // sides) and still record the session below.
-    tokio::select! {
-        _ = c2s => {},
-        _ = s2c => {},
-        _ = shutdown.cancelled() => {},
-    }
-
-    let duration_ms = start.elapsed().as_millis() as u64;
-    let mut session = WebSocketSession::new(tx_id);
-    session.id = session_id;
-    session.extensions = extensions;
-    session.duration_ms = duration_ms;
-    session.messages = match Arc::try_unwrap(messages) {
-        Ok(mutex) => mutex.into_inner(),
-        Err(arc) => arc.lock().await.clone(),
-    };
-    session.close_code = match Arc::try_unwrap(close_code) {
-        Ok(mutex) => mutex.into_inner(),
-        Err(arc) => *arc.lock().await,
-    };
-    session.violations = match Arc::try_unwrap(violations) {
-        Ok(mutex) => mutex.into_inner(),
-        Err(arc) => arc.lock().await.clone(),
-    };
-
-    if let Err(e) = captures.write_websocket_session(session).await {
-        error!("failed to write websocket session: {}", e);
-    }
-}
-
-fn message_to_info(
-    msg: &tokio_tungstenite::tungstenite::Message,
-    direction: crate::websocket_session::MessageDirection,
-) -> crate::websocket_session::WebSocketMessageInfo {
-    use crate::websocket_session::WebSocketMessageInfo;
-
-    // These numbers are the wire format, not tungstenite's enum discriminants,
-    // and this match is the only place the mapping is written down -- so it is
-    // the opcode table, restated in Rust.
-    //
-    // Quoted as one extract rather than per arm, for a plain reason: "%x9 denotes
-    // a ping" is eighteen characters and the extractor's floor is twenty. Ping and
-    // pong cannot be quoted alone, so the list is quoted whole and each arm reads
-    // its own line out of it.
-    //
-    // cite(RFC 6455 § 5.2): "%x0 denotes a continuation frame * %x1 denotes a text frame * %x2 denotes a binary frame * %x3-7 are reserved for further non-control frames * %x8 denotes a connection close * %x9 denotes a ping * %xA denotes a pong"
-    let (opcode, payload_length, fin, rsv, masked) = match msg {
-        // Assembled messages: tungstenite has already defragmented, so FIN is
-        // implicitly true and the RSV and MASK bits are not available. `None`
-        // is the honest record for the mask -- an assembled message has no
-        // header, and *not recorded* is not *not masked*, which is the value
-        // RFC 6455 § 5.1 makes a client's defect.
-        tokio_tungstenite::tungstenite::Message::Text(s) => (1, s.len() as u64, true, 0u8, None),
-        tokio_tungstenite::tungstenite::Message::Binary(b) => (2, b.len() as u64, true, 0, None),
-        tokio_tungstenite::tungstenite::Message::Ping(b) => (9, b.len() as u64, true, 0, None),
-        tokio_tungstenite::tungstenite::Message::Pong(b) => (10, b.len() as u64, true, 0, None),
-        tokio_tungstenite::tungstenite::Message::Close(frame) => {
-            let len = frame
-                .as_ref()
-                .map(|f| 2 + f.reason.len() as u64)
-                .unwrap_or(0);
-            (8, len, true, 0, None)
-        }
-        // Raw Frame variant: the header is there, so FIN, the RSV bits and the
-        // MASK bit are all read from it. `is_masked` is the bit rather than the
-        // key: § 5.2 defines the bit as *whether the payload data is masked*,
-        // and a key is present exactly when it is set.
-        tokio_tungstenite::tungstenite::Message::Frame(f) => {
-            let hdr = f.header();
-            let rsv_bits = ((hdr.rsv1 as u8) << 2) | ((hdr.rsv2 as u8) << 1) | (hdr.rsv3 as u8);
-            (
-                u8::from(hdr.opcode),
-                f.payload().len() as u64,
-                hdr.is_final,
-                rsv_bits,
-                Some(hdr.mask.is_some()),
-            )
-        }
-    };
-    WebSocketMessageInfo {
-        direction,
-        opcode,
-        payload_length,
-        fin,
-        rsv,
-        masked,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::capture::CaptureWriter;
-    use crate::proxy::test_support::make_shared_with_cfg;
-    use bytes::Bytes;
-    use http_body_util::{BodyExt, Full};
-    use hyper::{Method, Request};
-    use rstest::rstest;
-    use std::sync::Arc as StdArc;
-    use tokio::time::Instant;
-    use uuid::Uuid;
-
-    /// The MASK bit reaches the event only from the raw frame path, and an
-    /// assembled message records `None` rather than `false`.
-    ///
-    /// The distinction is the whole of `websocket_frame_masking`'s
-    /// decline: `false` is RFC 6455 § 5.1's finding against a client, and an
-    /// assembled `Text` has no header to have read it from.
-    #[test]
-    fn message_to_info_records_the_mask_bit_only_where_a_header_carried_one() {
-        use crate::websocket_session::MessageDirection;
-        use tokio_tungstenite::tungstenite::protocol::frame::{
-            coding::{Data, OpCode},
-            Frame, FrameHeader,
-        };
-        use tokio_tungstenite::tungstenite::Message;
-
-        let assembled = message_to_info(&Message::text("hi"), MessageDirection::Client);
-        assert_eq!(assembled.masked, None, "an assembled message has no header");
-
-        let header = FrameHeader {
-            is_final: true,
-            opcode: OpCode::Data(Data::Text),
-            mask: Some([1, 2, 3, 4]),
-            ..Default::default()
-        };
-        let masked = message_to_info(
-            &Message::Frame(Frame::from_payload(header, b"hi"[..].into())),
-            MessageDirection::Client,
-        );
-        assert_eq!(masked.masked, Some(true));
-
-        let header = FrameHeader {
-            is_final: true,
-            opcode: OpCode::Data(Data::Text),
-            mask: None,
-            ..Default::default()
-        };
-        let unmasked = message_to_info(
-            &Message::Frame(Frame::from_payload(header, b"hi"[..].into())),
-            MessageDirection::Client,
-        );
-        assert_eq!(unmasked.masked, Some(false));
-    }
-
-    /// The negotiation stops at the relay: the field is removed however many
-    /// lines carried it, and nothing else moves. The frames of an accepted
-    /// extension would be unreadable to tungstenite in the middle, so relaying
-    /// the offer or the acceptance authorised sessions the relay then killed
-    /// (RULECITES P48).
-    #[test]
-    fn extension_negotiation_is_stripped_and_nothing_else_is() {
-        use hyper::header::{HeaderName, HeaderValue};
-        let name = HeaderName::from_static("sec-websocket-extensions");
-
-        let mut headers = hyper::HeaderMap::new();
-        headers.insert(
-            hyper::header::UPGRADE,
-            HeaderValue::from_static("websocket"),
-        );
-        headers.insert(
-            HeaderName::from_static("sec-websocket-key"),
-            HeaderValue::from_static("dGhlIHNhbXBsZSBub25jZQ=="),
-        );
-        headers.append(name.clone(), HeaderValue::from_static("permessage-deflate"));
-        headers.append(name.clone(), HeaderValue::from_static("bbf-usp-protocol"));
-
-        let out = without_extension_negotiation(&headers);
-        assert!(!out.contains_key(&name), "both field lines are gone");
-        assert_eq!(out.len(), 2, "everything else survives");
-        assert!(out.contains_key(hyper::header::UPGRADE));
-        assert!(out.contains_key("sec-websocket-key"));
-    }
-
-    /// The one line that turns a `101` into the fact a frame rule reads.
-    ///
-    /// The absent case is the load-bearing one: it says *the server accepted
-    /// nothing*, which is what licenses `websocket_frame_rsv_bits` to
-    /// report a reserved bit at all. The `obs-text` case is the mirror — a
-    /// field that is there and unreadable must not become "accepted nothing",
-    /// or an unreadable handshake would start licensing findings.
-    #[test]
-    fn accepted_extensions_reads_the_response_field() {
-        use crate::protocol_event::NegotiatedExtensions;
-        use hyper::header::{HeaderName, HeaderValue};
-
-        let name = HeaderName::from_static("sec-websocket-extensions");
-
-        let mut none = hyper::HeaderMap::new();
-        none.insert(
-            hyper::header::UPGRADE,
-            HeaderValue::from_static("websocket"),
-        );
-        assert_eq!(
-            accepted_extensions(&none),
-            NegotiatedExtensions::NoneAccepted
-        );
-
-        let mut one = hyper::HeaderMap::new();
-        one.insert(name.clone(), HeaderValue::from_static("permessage-deflate"));
-        assert_eq!(
-            accepted_extensions(&one),
-            NegotiatedExtensions::Accepted("permessage-deflate".into())
-        );
-
-        // Several field lines in one section are one value.
-        let mut two = hyper::HeaderMap::new();
-        two.append(name.clone(), HeaderValue::from_static("foo"));
-        two.append(name.clone(), HeaderValue::from_static("bar; baz=2"));
-        assert_eq!(
-            accepted_extensions(&two),
-            NegotiatedExtensions::Accepted("foo,bar; baz=2".into())
-        );
-
-        let mut obs = hyper::HeaderMap::new();
-        obs.insert(
-            name,
-            HeaderValue::from_bytes(&[0xff]).expect("obs-text is a field-content"),
-        );
-        assert!(
-            matches!(accepted_extensions(&obs), NegotiatedExtensions::Accepted(_)),
-            "an unreadable value is still a field the server sent"
-        );
-    }
-
-    fn test_pe_pipeline() -> ProtocolEventPipeline {
-        let cfg = crate::config::Config::default();
-        let engine = StdArc::new(crate::engine::PreparedEngine::new(&cfg).unwrap());
-        ProtocolEventPipeline::new(
-            engine,
-            StdArc::new(crate::protocol_event_store::ProtocolEventStore::new(
-                300, 100,
-            )),
-        )
-    }
-
-    /// A trust-store-free client config for the `ws://` (non-TLS) tests below —
-    /// they never reach the HTTPS branch, so an empty root store is fine.
-    fn test_tls_config() -> StdArc<rustls::ClientConfig> {
-        StdArc::new(
-            rustls::ClientConfig::builder()
-                .with_root_certificates(rustls::RootCertStore::empty())
-                .with_no_client_auth(),
-        )
-    }
-
-    /// Whether a captured transaction's request headers (serialized as ordered
-    /// `[name, value]` pairs) contain `name`.
-    fn captured_request_has_header(v: &serde_json::Value, name: &str) -> bool {
-        v["request"]["headers"]
-            .as_array()
-            .map(|pairs| pairs.iter().any(|p| p[0] == name))
-            .unwrap_or(false)
-    }
-
-    #[test]
-    fn is_websocket_upgrade_detects_valid_upgrade() {
-        let req = Request::builder()
-            .method("GET")
-            .uri("http://example.com/ws")
-            .header("connection", "Upgrade")
-            .header("upgrade", "websocket")
-            .body(Full::new(Bytes::new()).boxed())
-            .unwrap();
-        assert!(is_websocket_upgrade(&req));
-    }
-
-    #[test]
-    fn is_websocket_upgrade_case_insensitive() {
-        let req = Request::builder()
-            .method("GET")
-            .uri("http://example.com/ws")
-            .header("connection", "upgrade")
-            .header("upgrade", "WebSocket")
-            .body(Full::new(Bytes::new()).boxed())
-            .unwrap();
-        assert!(is_websocket_upgrade(&req));
-    }
-
-    #[rstest]
-    #[case(Some("keep-alive"), Some("websocket"), false)]
-    #[case(Some("Upgrade"), None, false)]
-    #[case(None, Some("websocket"), false)]
-    #[case(None, None, false)]
-    #[case(Some("Upgrade"), Some("h2c"), false)]
-    // Connection contains "upgrade" only as a substring of another token, and
-    // RFC 9110 §7.6.1 makes Connection a list of tokens, not a string to search.
-    // The sentence is cited where the splitting happens, in `parse_connection_tokens`.
-    #[case(Some("super-upgrade"), Some("websocket"), false)]
-    #[case(Some("upgrades"), Some("websocket"), false)]
-    fn is_websocket_upgrade_negative(
-        #[case] connection: Option<&str>,
-        #[case] upgrade: Option<&str>,
-        #[case] expected: bool,
-    ) {
-        let mut builder = Request::builder()
-            .method("GET")
-            .uri("http://example.com/ws");
-        if let Some(c) = connection {
-            builder = builder.header("connection", c);
-        }
-        if let Some(u) = upgrade {
-            builder = builder.header("upgrade", u);
-        }
-        let req = builder.body(Full::new(Bytes::new()).boxed()).unwrap();
-        assert_eq!(is_websocket_upgrade(&req), expected);
-    }
-
-    #[test]
-    fn message_to_info_text() {
-        use crate::websocket_session::MessageDirection;
-        let msg = tokio_tungstenite::tungstenite::Message::Text("hello".into());
-        let info = message_to_info(&msg, MessageDirection::Client);
-        assert_eq!(info.opcode, 1);
-        assert_eq!(info.payload_length, 5);
-        assert_eq!(info.direction, MessageDirection::Client);
-    }
-
-    #[test]
-    fn message_to_info_binary() {
-        use crate::websocket_session::MessageDirection;
-        let msg = tokio_tungstenite::tungstenite::Message::Binary(vec![1, 2, 3].into());
-        let info = message_to_info(&msg, MessageDirection::Server);
-        assert_eq!(info.opcode, 2);
-        assert_eq!(info.payload_length, 3);
-        assert_eq!(info.direction, MessageDirection::Server);
-    }
-
-    #[test]
-    fn message_to_info_ping_pong() {
-        use crate::websocket_session::MessageDirection;
-        let ping = tokio_tungstenite::tungstenite::Message::Ping(vec![0; 4].into());
-        let info = message_to_info(&ping, MessageDirection::Client);
-        assert_eq!(info.opcode, 9);
-        assert_eq!(info.payload_length, 4);
-
-        let pong = tokio_tungstenite::tungstenite::Message::Pong(vec![0; 2].into());
-        let info = message_to_info(&pong, MessageDirection::Server);
-        assert_eq!(info.opcode, 10);
-        assert_eq!(info.payload_length, 2);
-    }
-
-    #[test]
-    fn message_to_info_close_with_frame() {
-        use crate::websocket_session::MessageDirection;
-        use tokio_tungstenite::tungstenite::protocol::CloseFrame;
-        let msg = tokio_tungstenite::tungstenite::Message::Close(Some(CloseFrame {
-            code: tokio_tungstenite::tungstenite::protocol::frame::coding::CloseCode::Normal,
-            reason: "bye".into(),
-        }));
-        let info = message_to_info(&msg, MessageDirection::Client);
-        assert_eq!(info.opcode, 8);
-        // 2 bytes for code + 3 bytes for "bye"
-        assert_eq!(info.payload_length, 5);
-    }
-
-    #[test]
-    fn message_to_info_close_without_frame() {
-        use crate::websocket_session::MessageDirection;
-        let msg = tokio_tungstenite::tungstenite::Message::Close(None);
-        let info = message_to_info(&msg, MessageDirection::Server);
-        assert_eq!(info.opcode, 8);
-        assert_eq!(info.payload_length, 0);
-    }
-
-    #[tokio::test]
-    async fn relay_websocket_relays_messages_and_captures_session() -> anyhow::Result<()> {
-        use futures_util::{SinkExt, StreamExt};
-        use tokio_tungstenite::tungstenite::protocol::Role;
-
-        // Create two duplex pairs to simulate client<->proxy and proxy<->server
-        let (client_side, proxy_client_side) = tokio::io::duplex(4096);
-        let (proxy_server_side, server_side) = tokio::io::duplex(4096);
-
-        let tx_id = Uuid::new_v4();
-        let tmp = std::env::temp_dir().join(format!("lint_ws_relay_test_{}.jsonl", Uuid::new_v4()));
-        let p = tmp.to_str().unwrap().to_string();
-        let cw = CaptureWriter::new(p.clone(), false).await?;
-
-        // Spawn the relay
-        let cw_clone = cw.clone();
-        let relay_handle = tokio::spawn(async move {
-            relay_websocket(
-                proxy_client_side,
-                proxy_server_side,
-                tx_id,
-                cw_clone,
-                uuid::Uuid::new_v4(),
-                crate::protocol_event::NegotiatedExtensions::NoneAccepted,
-                test_pe_pipeline(),
-                tokio_util::sync::CancellationToken::new(),
-            )
-            .await;
-        });
-
-        // Client side: wrap in WebSocket (client role)
-        let mut client_ws =
-            tokio_tungstenite::WebSocketStream::from_raw_socket(client_side, Role::Client, None)
-                .await;
-
-        // Server side: wrap in WebSocket (server role)
-        let mut server_ws =
-            tokio_tungstenite::WebSocketStream::from_raw_socket(server_side, Role::Server, None)
-                .await;
-
-        // Client sends a text message
-        client_ws
-            .send(tokio_tungstenite::tungstenite::Message::Text(
-                "hello".into(),
-            ))
-            .await?;
-
-        // Server should receive it
-        let msg = server_ws.next().await.unwrap()?;
-        assert_eq!(
-            msg,
-            tokio_tungstenite::tungstenite::Message::Text("hello".into())
-        );
-
-        // Server sends a response
-        server_ws
-            .send(tokio_tungstenite::tungstenite::Message::Text(
-                "world".into(),
-            ))
-            .await?;
-
-        // Client should receive it
-        let msg = client_ws.next().await.unwrap()?;
-        assert_eq!(
-            msg,
-            tokio_tungstenite::tungstenite::Message::Text("world".into())
-        );
-
-        // Client sends close
-        client_ws
-            .send(tokio_tungstenite::tungstenite::Message::Close(Some(
-                tokio_tungstenite::tungstenite::protocol::CloseFrame {
-                    code:
-                        tokio_tungstenite::tungstenite::protocol::frame::coding::CloseCode::Normal,
-                    reason: "done".into(),
-                },
-            )))
-            .await?;
-
-        // Server receives close and sends close back
-        let msg = server_ws.next().await.unwrap()?;
-        assert!(matches!(
-            msg,
-            tokio_tungstenite::tungstenite::Message::Close(_)
-        ));
-        server_ws.close(None).await.ok();
-
-        // Close client side
-        client_ws.close(None).await.ok();
-
-        // Wait for relay to finish
-        tokio::time::timeout(std::time::Duration::from_secs(5), relay_handle)
-            .await
-            .expect("relay did not finish")
-            .expect("relay panicked");
-
-        // Flush captures
-        cw.flush().await?;
-
-        // Read the capture file and verify the WebSocket session was written
-        let content = tokio::fs::read_to_string(&p).await?;
-        assert!(!content.is_empty(), "capture file should not be empty");
-        let session: serde_json::Value = serde_json::from_str(content.trim())?;
-        assert_eq!(session["type"].as_str(), Some("websocket_session"));
-        assert_eq!(
-            session["transaction_id"].as_str(),
-            Some(tx_id.to_string().as_str())
-        );
-        let messages = session["messages"].as_array().unwrap();
-        assert!(messages.len() >= 2, "should have at least 2 messages");
-        // First message should be client text
-        assert_eq!(messages[0]["direction"].as_str(), Some("client"));
-        assert_eq!(messages[0]["opcode"].as_u64(), Some(1));
-        // Second message should be server text
-        assert_eq!(messages[1]["direction"].as_str(), Some("server"));
-        assert_eq!(messages[1]["opcode"].as_u64(), Some(1));
-        // Should have a close code
-        assert_eq!(session["close_code"].as_u64(), Some(1000));
-
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn connect_upstream_for_upgrade_fails_without_host() {
-        let uri: Uri = "/no-host".parse().unwrap();
-        let result =
-            connect_upstream_for_upgrade(&uri, &hyper::http::uri::Scheme::HTTP, &test_tls_config())
-                .await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn connect_upstream_for_upgrade_fails_with_closed_port() {
-        // pick a port that's not listening
-        let l = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
-        let port = l.local_addr().unwrap().port();
-        drop(l);
-        let uri: Uri = format!("http://127.0.0.1:{}/ws", port).parse().unwrap();
-        let result =
-            connect_upstream_for_upgrade(&uri, &hyper::http::uri::Scheme::HTTP, &test_tls_config())
-                .await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn handle_websocket_upgrade_upstream_connect_error() -> anyhow::Result<()> {
-        // Test that handle_websocket_upgrade returns 502 when upstream is unreachable
-        let cfg = StdArc::new(crate::config::Config::default());
-        let (shared, tmp, cw) = make_shared_with_cfg(cfg, None).await?;
-
-        // Build a request targeting a closed port
-        let l = std::net::TcpListener::bind("127.0.0.1:0")?;
-        let port = l.local_addr()?.port();
-        drop(l);
-
-        let uri: Uri = format!("http://127.0.0.1:{}/ws", port).parse()?;
-        let upstream_req = Request::builder()
-            .method("GET")
-            .uri(uri.clone())
-            .header("connection", "Upgrade")
-            .header("upgrade", "websocket")
-            .body(Full::new(Bytes::new()))?;
-
-        // Create a fake OnUpgrade that will never complete
-        let fake_on_upgrade = hyper::upgrade::on(
-            Request::builder()
-                .method("GET")
-                .uri("http://fake/")
-                .body(Full::new(Bytes::new()).boxed())
-                .unwrap(),
-        );
-
-        let conn_metadata = StdArc::new(crate::connection::ConnectionMetadata::new(
-            "127.0.0.1:12345".parse()?,
-        ));
-        let started = Instant::now();
-        let client_id =
-            crate::state::ClientIdentifier::new("127.0.0.1".parse().unwrap(), "test".to_string());
-        let mut req_headers = hyper::HeaderMap::new();
-        req_headers.insert("x-test", "1".parse()?);
-
-        let resp = handle_websocket_upgrade(
-            upstream_req,
-            fake_on_upgrade,
-            &uri,
-            &hyper::http::uri::Scheme::HTTP,
-            &started,
-            &client_id,
-            &Method::GET,
-            &uri.to_string(),
-            &req_headers,
-            "HTTP/1.1",
-            Bytes::new(),
-            None,
-            shared,
-            conn_metadata,
-        )
-        .await?;
-
-        assert_eq!(resp.status().as_u16(), 502);
-
-        // The failed handshake is now captured rather than silently dropped,
-        // preserving the request headers.
-        cw.flush().await?;
-        let content = tokio::fs::read_to_string(&tmp).await?;
-        let v: serde_json::Value = serde_json::from_str(content.trim())?;
-        assert_eq!(v["response"]["status"].as_u64(), Some(502));
-        assert!(
-            captured_request_has_header(&v, "x-test"),
-            "captured request should preserve request headers"
-        );
-
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn handle_websocket_upgrade_non_101_response() -> anyhow::Result<()> {
-        // Start a plain HTTP server that returns 400 for upgrade requests
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await?;
-        let port = listener.local_addr()?.port();
-
-        let server_task = tokio::spawn(async move {
-            if let Ok((socket, _)) = listener.accept().await {
-                let mut buf = [0u8; 4096];
-                let _ = socket.readable().await;
-                let _ = socket.try_read(&mut buf);
-                let resp = b"HTTP/1.1 400 Bad Request\r\nContent-Length: 11\r\n\r\nBad Request";
-                let _ = socket.try_write(resp);
-            }
-        });
-
-        let cfg = StdArc::new(crate::config::Config::default());
-        let (shared, tmp, _cw) = make_shared_with_cfg(cfg, None).await?;
-
-        let uri: Uri = format!("http://127.0.0.1:{}/ws", port).parse()?;
-        let upstream_req = Request::builder()
-            .method("GET")
-            .uri(uri.clone())
-            .header("connection", "Upgrade")
-            .header("upgrade", "websocket")
-            .body(Full::new(Bytes::new()))?;
-
-        let fake_on_upgrade = hyper::upgrade::on(
-            Request::builder()
-                .method("GET")
-                .uri("http://fake/")
-                .body(Full::new(Bytes::new()).boxed())
-                .unwrap(),
-        );
-
-        let conn_metadata = StdArc::new(crate::connection::ConnectionMetadata::new(
-            "127.0.0.1:12345".parse()?,
-        ));
-        let started = Instant::now();
-        let client_id =
-            crate::state::ClientIdentifier::new("127.0.0.1".parse().unwrap(), "test".to_string());
-
-        let resp = handle_websocket_upgrade(
-            upstream_req,
-            fake_on_upgrade,
-            &uri,
-            &hyper::http::uri::Scheme::HTTP,
-            &started,
-            &client_id,
-            &Method::GET,
-            &uri.to_string(),
-            &hyper::HeaderMap::new(),
-            "HTTP/1.1",
-            Bytes::new(),
-            None,
-            shared,
-            conn_metadata,
-        )
-        .await?;
-
-        // Server returned 400, so proxy should forward it
-        assert_eq!(resp.status().as_u16(), 400);
-
-        let _ = server_task.await;
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn handle_websocket_upgrade_non_101_streams_response() -> anyhow::Result<()> {
-        // Plain HTTP server rejecting the upgrade with a body larger than the
-        // old `max_body_bytes` guard — it must now stream through (no 502).
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await?;
-        let port = listener.local_addr()?.port();
-
-        let server_task = tokio::spawn(async move {
-            if let Ok((socket, _)) = listener.accept().await {
-                let mut buf = [0u8; 4096];
-                let _ = socket.readable().await;
-                let _ = socket.try_read(&mut buf);
-                let resp = b"HTTP/1.1 400 Bad Request\r\nContent-Length: 11\r\n\r\nBad Request";
-                let _ = socket.try_write(resp);
-            }
-        });
-
-        let mut cfg = crate::config::Config::default();
-        cfg.general.max_body_bytes = 4;
-        let (shared, tmp, _cw) = make_shared_with_cfg(StdArc::new(cfg), None).await?;
-
-        let uri: Uri = format!("http://127.0.0.1:{}/ws", port).parse()?;
-        let upstream_req = Request::builder()
-            .method("GET")
-            .uri(uri.clone())
-            .header("connection", "Upgrade")
-            .header("upgrade", "websocket")
-            .body(Full::new(Bytes::new()))?;
-
-        let fake_on_upgrade = hyper::upgrade::on(
-            Request::builder()
-                .method("GET")
-                .uri("http://fake/")
-                .body(Full::new(Bytes::new()).boxed())
-                .unwrap(),
-        );
-
-        let conn_metadata = StdArc::new(crate::connection::ConnectionMetadata::new(
-            "127.0.0.1:12345".parse()?,
-        ));
-        let started = Instant::now();
-        let client_id =
-            crate::state::ClientIdentifier::new("127.0.0.1".parse().unwrap(), "test".to_string());
-
-        let resp = handle_websocket_upgrade(
-            upstream_req,
-            fake_on_upgrade,
-            &uri,
-            &hyper::http::uri::Scheme::HTTP,
-            &started,
-            &client_id,
-            &Method::GET,
-            &uri.to_string(),
-            &hyper::HeaderMap::new(),
-            "HTTP/1.1",
-            Bytes::new(),
-            None,
-            shared,
-            conn_metadata,
-        )
-        .await?;
-
-        // The upstream's non-101 status and body stream through unchanged — no 502.
-        assert_eq!(resp.status().as_u16(), 400);
-        let body = resp.into_body().collect().await.unwrap().to_bytes();
-        assert_eq!(&body[..], b"Bad Request");
-
-        // The handshake transaction is captured with the upstream's real status.
-        _cw.flush().await?;
-        let content = tokio::fs::read_to_string(&tmp).await?;
-        let v: serde_json::Value = serde_json::from_str(content.trim())?;
-        assert_eq!(v["response"]["status"].as_u64(), Some(400));
-        assert_eq!(v["response_body_over_limit"].as_bool(), Some(false));
-
-        let _ = server_task.await;
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn relay_websocket_server_initiated_close() -> anyhow::Result<()> {
-        // Test the s2c direction: server sends close first
-        use futures_util::{SinkExt, StreamExt};
-        use tokio_tungstenite::tungstenite::protocol::Role;
-
-        let (client_side, proxy_client_side) = tokio::io::duplex(4096);
-        let (proxy_server_side, server_side) = tokio::io::duplex(4096);
-
-        let tx_id = Uuid::new_v4();
-        let tmp = std::env::temp_dir().join(format!("lint_ws_s2c_test_{}.jsonl", Uuid::new_v4()));
-        let p = tmp.to_str().unwrap().to_string();
-        let cw = CaptureWriter::new(p.clone(), false).await?;
-
-        let cw_clone = cw.clone();
-        let relay_handle = tokio::spawn(async move {
-            relay_websocket(
-                proxy_client_side,
-                proxy_server_side,
-                tx_id,
-                cw_clone,
-                uuid::Uuid::new_v4(),
-                crate::protocol_event::NegotiatedExtensions::NoneAccepted,
-                test_pe_pipeline(),
-                tokio_util::sync::CancellationToken::new(),
-            )
-            .await;
-        });
-
-        let mut client_ws =
-            tokio_tungstenite::WebSocketStream::from_raw_socket(client_side, Role::Client, None)
-                .await;
-
-        let mut server_ws =
-            tokio_tungstenite::WebSocketStream::from_raw_socket(server_side, Role::Server, None)
-                .await;
-
-        // Server sends a message first
-        server_ws
-            .send(tokio_tungstenite::tungstenite::Message::Text(
-                "server-msg".into(),
-            ))
-            .await?;
-        let msg = client_ws.next().await.unwrap()?;
-        assert_eq!(
-            msg,
-            tokio_tungstenite::tungstenite::Message::Text("server-msg".into())
-        );
-
-        // Server initiates close
-        server_ws
-            .send(tokio_tungstenite::tungstenite::Message::Close(Some(
-                tokio_tungstenite::tungstenite::protocol::CloseFrame {
-                    code:
-                        tokio_tungstenite::tungstenite::protocol::frame::coding::CloseCode::Normal,
-                    reason: "server-done".into(),
-                },
-            )))
-            .await?;
-
-        // Client receives close and responds
-        let msg = client_ws.next().await.unwrap()?;
-        assert!(matches!(
-            msg,
-            tokio_tungstenite::tungstenite::Message::Close(_)
-        ));
-        // Send close response then drop to end the c2s stream
-        client_ws
-            .send(tokio_tungstenite::tungstenite::Message::Close(None))
-            .await
-            .ok();
-        drop(client_ws);
-        drop(server_ws);
-
-        tokio::time::timeout(std::time::Duration::from_secs(5), relay_handle)
-            .await
-            .expect("relay did not finish")
-            .expect("relay panicked");
-
-        cw.flush().await?;
-
-        let content = tokio::fs::read_to_string(&p).await?;
-        let session: serde_json::Value = serde_json::from_str(content.trim())?;
-        assert_eq!(session["type"].as_str(), Some("websocket_session"));
-        // Close was from server direction
-        assert_eq!(session["close_code"].as_u64(), Some(1000));
-        let messages = session["messages"].as_array().unwrap();
-        // Should have server text + server close + possibly client close
-        assert!(messages.len() >= 2);
-
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn relay_websocket_abrupt_disconnect() -> anyhow::Result<()> {
-        // Test error path: client disconnects abruptly
-        let (client_side, proxy_client_side) = tokio::io::duplex(4096);
-        let (proxy_server_side, server_side) = tokio::io::duplex(4096);
-
-        let tx_id = Uuid::new_v4();
-        let tmp =
-            std::env::temp_dir().join(format!("lint_ws_abrupt_test_{}.jsonl", Uuid::new_v4()));
-        let p = tmp.to_str().unwrap().to_string();
-        let cw = CaptureWriter::new(p.clone(), false).await?;
-
-        let cw_clone = cw.clone();
-        let relay_handle = tokio::spawn(async move {
-            relay_websocket(
-                proxy_client_side,
-                proxy_server_side,
-                tx_id,
-                cw_clone,
-                uuid::Uuid::new_v4(),
-                crate::protocol_event::NegotiatedExtensions::NoneAccepted,
-                test_pe_pipeline(),
-                tokio_util::sync::CancellationToken::new(),
-            )
-            .await;
-        });
-
-        // Drop client and server immediately to simulate abrupt disconnect
-        drop(client_side);
-        drop(server_side);
-
-        tokio::time::timeout(std::time::Duration::from_secs(5), relay_handle)
-            .await
-            .expect("relay did not finish")
-            .expect("relay panicked");
-
-        cw.flush().await?;
-
-        // Should still write a session record (with empty messages)
-        let content = tokio::fs::read_to_string(&p).await?;
-        let session: serde_json::Value = serde_json::from_str(content.trim())?;
-        assert_eq!(session["type"].as_str(), Some("websocket_session"));
-
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn relay_websocket_closes_on_shutdown() -> anyhow::Result<()> {
-        // Both peers stay connected (no frames, no close), so the relay only
-        // ends when the shutdown token is cancelled — proving it observes it
-        // and closes gracefully rather than lingering to the drain timeout.
-        let (client_side, proxy_client_side) = tokio::io::duplex(4096);
-        let (proxy_server_side, server_side) = tokio::io::duplex(4096);
-        // Keep the peer ends alive so the relay's reads stay pending (not EOF).
-        let _client_side = client_side;
-        let _server_side = server_side;
-
-        let tx_id = Uuid::new_v4();
-        let tmp =
-            std::env::temp_dir().join(format!("lint_ws_shutdown_test_{}.jsonl", Uuid::new_v4()));
-        let p = tmp.to_str().unwrap().to_string();
-        let cw = CaptureWriter::new(p.clone(), false).await?;
-
-        let shutdown = tokio_util::sync::CancellationToken::new();
-        let cw_clone = cw.clone();
-        let shutdown_relay = shutdown.clone();
-        let relay_handle = tokio::spawn(async move {
-            relay_websocket(
-                proxy_client_side,
-                proxy_server_side,
-                tx_id,
-                cw_clone,
-                uuid::Uuid::new_v4(),
-                crate::protocol_event::NegotiatedExtensions::NoneAccepted,
-                test_pe_pipeline(),
-                shutdown_relay,
-            )
-            .await;
-        });
-
-        // Let the relay start and block on its idle reads, then signal shutdown.
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        shutdown.cancel();
-
-        tokio::time::timeout(std::time::Duration::from_secs(5), relay_handle)
-            .await
-            .expect("relay did not stop on shutdown")
-            .expect("relay panicked");
-
-        // The session is still recorded on the shutdown path.
-        cw.flush().await?;
-        let content = tokio::fs::read_to_string(&p).await?;
-        let session: serde_json::Value = serde_json::from_str(content.trim())?;
-        assert_eq!(session["type"].as_str(), Some("websocket_session"));
-
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-
-    #[test]
-    fn message_to_info_frame_variant() {
-        use crate::websocket_session::MessageDirection;
-        use tokio_tungstenite::tungstenite::protocol::frame::coding::OpCode;
-        use tokio_tungstenite::tungstenite::protocol::frame::{Frame, FrameHeader};
-        let header = FrameHeader {
-            is_final: true,
-            rsv1: false,
-            rsv2: false,
-            rsv3: false,
-            opcode: OpCode::Data(
-                tokio_tungstenite::tungstenite::protocol::frame::coding::Data::Text,
-            ),
-            mask: None,
-        };
-        let frame = Frame::from_payload(header, vec![b'h', b'i'].into());
-        let msg = tokio_tungstenite::tungstenite::Message::Frame(frame);
-        let info = message_to_info(&msg, MessageDirection::Client);
-        assert_eq!(info.opcode, 1); // Text opcode
-        assert_eq!(info.payload_length, 2);
-    }
-
-    /// The dependency on the capture path, pinned: a reserved opcode never
-    /// reaches `message_to_info`, because the frame reader refuses it in the
-    /// header parser before any of the rest of the header is used. Two ranges,
-    /// and everything the document defines still parses.
-    ///
-    /// `websocket_frame_opcode_sequence` reports reserved opcodes, and
-    /// this is why its `description()` says those findings arrive through
-    /// `lint` over a capture file some other tool wrote rather than off this
-    /// proxy's own relay. When this test fails, that paragraph is what has gone
-    /// stale.
-    #[test]
-    fn the_frame_reader_refuses_a_reserved_opcode_before_the_capture_path() {
-        use std::io::Cursor;
-        use tokio_tungstenite::tungstenite::protocol::frame::FrameHeader;
-
-        // First byte: FIN set, no reserved bits, opcode in the low nibble.
-        // Second byte: unmasked, payload length 0.
-        for reserved in [0x3u8, 0x4, 0x5, 0x6, 0x7, 0xB, 0xC, 0xD, 0xE, 0xF] {
-            let head = [0x80 | reserved, 0x00];
-            assert!(
-                FrameHeader::parse(&mut Cursor::new(&head)).is_err(),
-                "opcode {reserved:#x} must not reach a capture"
-            );
-        }
-        for defined in [0x0u8, 0x1, 0x2, 0x8, 0x9, 0xA] {
-            let head = [0x80 | defined, 0x00];
-            let parsed = FrameHeader::parse(&mut Cursor::new(&head))
-                .expect("a defined opcode parses")
-                .expect("the two-byte head is complete");
-            assert_eq!(u8::from(parsed.0.opcode), defined);
-        }
-    }
-
-    #[tokio::test]
-    async fn connect_upstream_for_upgrade_plain_tcp_success() -> anyhow::Result<()> {
-        // Start a simple HTTP server that accepts connections
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await?;
-        let port = listener.local_addr()?.port();
-
-        let server_task = tokio::spawn(async move {
-            if let Ok((socket, _)) = listener.accept().await {
-                let mut buf = [0u8; 4096];
-                let _ = socket.readable().await;
-                let _ = socket.try_read(&mut buf);
-                let resp = b"HTTP/1.1 200 OK\r\nContent-Length: 2\r\n\r\nok";
-                let _ = socket.try_write(resp);
-            }
-        });
-
-        let uri: Uri = format!("http://127.0.0.1:{}/ws", port).parse()?;
-        let (mut sender, _handle) =
-            connect_upstream_for_upgrade(&uri, &hyper::http::uri::Scheme::HTTP, &test_tls_config())
-                .await?;
-
-        // Verify we can send a request
-        let req = Request::builder()
-            .method("GET")
-            .uri(format!("http://127.0.0.1:{}/ws", port))
-            .body(Full::new(Bytes::new()))?;
-        let resp = sender.send_request(req).await?;
-        assert_eq!(resp.status().as_u16(), 200);
-
-        let _ = server_task.await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn handle_websocket_upgrade_send_request_error() -> anyhow::Result<()> {
-        // Start a server that accepts TCP connections then immediately closes them
-        // This will cause the hyper handshake to succeed but send_request to fail
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await?;
-        let port = listener.local_addr()?.port();
-
-        let server_task = tokio::spawn(async move {
-            // Accept connection then drop it immediately
-            if let Ok((socket, _)) = listener.accept().await {
-                drop(socket);
-            }
-        });
-
-        let cfg = StdArc::new(crate::config::Config::default());
-        let (shared, tmp, cw) = make_shared_with_cfg(cfg, None).await?;
-
-        let uri: Uri = format!("http://127.0.0.1:{}/ws", port).parse()?;
-        let upstream_req = Request::builder()
-            .method("GET")
-            .uri(uri.clone())
-            .header("connection", "Upgrade")
-            .header("upgrade", "websocket")
-            .body(Full::new(Bytes::new()))?;
-
-        let fake_on_upgrade = hyper::upgrade::on(
-            Request::builder()
-                .method("GET")
-                .uri("http://fake/")
-                .body(Full::new(Bytes::new()).boxed())
-                .unwrap(),
-        );
-
-        let conn_metadata = StdArc::new(crate::connection::ConnectionMetadata::new(
-            "127.0.0.1:12345".parse()?,
-        ));
-        let started = Instant::now();
-        let client_id =
-            crate::state::ClientIdentifier::new("127.0.0.1".parse().unwrap(), "test".to_string());
-        let mut req_headers = hyper::HeaderMap::new();
-        req_headers.insert("x-test", "1".parse()?);
-
-        let resp = handle_websocket_upgrade(
-            upstream_req,
-            fake_on_upgrade,
-            &uri,
-            &hyper::http::uri::Scheme::HTTP,
-            &started,
-            &client_id,
-            &Method::GET,
-            &uri.to_string(),
-            &req_headers,
-            "HTTP/1.1",
-            Bytes::new(),
-            None,
-            shared,
-            conn_metadata,
-        )
-        .await?;
-
-        // Server dropped connection, send_request should fail -> 502
-        assert_eq!(resp.status().as_u16(), 502);
-
-        // The failed handshake is captured rather than silently dropped.
-        cw.flush().await?;
-        let content = tokio::fs::read_to_string(&tmp).await?;
-        let v: serde_json::Value = serde_json::from_str(content.trim())?;
-        assert_eq!(v["response"]["status"].as_u64(), Some(502));
-        assert!(
-            captured_request_has_header(&v, "x-test"),
-            "captured request should preserve request headers"
-        );
-
-        let _ = server_task.await;
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn relay_websocket_binary_and_ping_messages() -> anyhow::Result<()> {
-        // Test relay with binary and ping/pong messages to cover more message_to_info paths
-        use futures_util::{SinkExt, StreamExt};
-        use tokio_tungstenite::tungstenite::protocol::Role;
-
-        let (client_side, proxy_client_side) = tokio::io::duplex(4096);
-        let (proxy_server_side, server_side) = tokio::io::duplex(4096);
-
-        let tx_id = Uuid::new_v4();
-        let tmp =
-            std::env::temp_dir().join(format!("lint_ws_binary_test_{}.jsonl", Uuid::new_v4()));
-        let p = tmp.to_str().unwrap().to_string();
-        let cw = CaptureWriter::new(p.clone(), false).await?;
-
-        let cw_clone = cw.clone();
-        let relay_handle = tokio::spawn(async move {
-            relay_websocket(
-                proxy_client_side,
-                proxy_server_side,
-                tx_id,
-                cw_clone,
-                uuid::Uuid::new_v4(),
-                crate::protocol_event::NegotiatedExtensions::NoneAccepted,
-                test_pe_pipeline(),
-                tokio_util::sync::CancellationToken::new(),
-            )
-            .await;
-        });
-
-        let mut client_ws =
-            tokio_tungstenite::WebSocketStream::from_raw_socket(client_side, Role::Client, None)
-                .await;
-
-        let mut server_ws =
-            tokio_tungstenite::WebSocketStream::from_raw_socket(server_side, Role::Server, None)
-                .await;
-
-        // Client sends binary
-        client_ws
-            .send(tokio_tungstenite::tungstenite::Message::Binary(
-                vec![1, 2, 3].into(),
-            ))
-            .await?;
-        let msg = server_ws.next().await.unwrap()?;
-        assert!(matches!(
-            msg,
-            tokio_tungstenite::tungstenite::Message::Binary(_)
-        ));
-
-        // Server sends binary back
-        server_ws
-            .send(tokio_tungstenite::tungstenite::Message::Binary(
-                vec![4, 5, 6].into(),
-            ))
-            .await?;
-        let msg = client_ws.next().await.unwrap()?;
-        assert!(matches!(
-            msg,
-            tokio_tungstenite::tungstenite::Message::Binary(_)
-        ));
-
-        // Client sends ping
-        client_ws
-            .send(tokio_tungstenite::tungstenite::Message::Ping(
-                vec![7, 8].into(),
-            ))
-            .await?;
-
-        // Server receives ping (may receive pong auto-response)
-        let msg = server_ws.next().await.unwrap()?;
-        assert!(
-            matches!(msg, tokio_tungstenite::tungstenite::Message::Ping(_))
-                || matches!(msg, tokio_tungstenite::tungstenite::Message::Pong(_))
-        );
-
-        // Close
-        client_ws
-            .send(tokio_tungstenite::tungstenite::Message::Close(None))
-            .await
-            .ok();
-        drop(client_ws);
-        drop(server_ws);
-
-        tokio::time::timeout(std::time::Duration::from_secs(5), relay_handle)
-            .await
-            .expect("relay did not finish")
-            .expect("relay panicked");
-
-        cw.flush().await?;
-
-        let content = tokio::fs::read_to_string(&p).await?;
-        let session: serde_json::Value = serde_json::from_str(content.trim())?;
-        let messages = session["messages"].as_array().unwrap();
-        // Should have binary c2s, binary s2c, ping, and close messages
-        assert!(messages.len() >= 3);
-
-        // Verify binary opcode (2) appears
-        assert!(messages.iter().any(|m| m["opcode"].as_u64() == Some(2)));
-
-        let _ = tokio::fs::remove_file(&tmp).await;
-        Ok(())
-    }
-}
